+++ src/Reports/Logs.php
@@ -1,37 +1,32 @@
 <?php
 
-namespace WombatDialer\Controllers\Edit\Live;
+namespace WombatDialer\Reports;
 
 use Illuminate\Support\Facades\Http;
 use WombatDialer\Controllers\Edit\Wombat;
 
-class Calls extends Wombat
+class Logs extends Wombat
 {
     //protected $path = '/edit/asterisk';
     use \WombatDialer\Concerns\LiveCallsandReportsTraits;
 
     /**
      * Perform API GET.
-     * Returns the list of all the Live calls.
+     * Returns the Log set for the set of runs.
      *
+     * @params  $items and $from and $id
      * @return json response
      */
-    public function liveCalls()
+    public function reportLogs($from, $items, $id)
     {
+        $this->path = '/reports/logs/';
+        $from = $from ? $from : null;
+        $items = $items ? $items : null;
+        $id = $id ? $id : null;
+        $this->query = ['from' => $from, 'items' => $items, 'id' => $id];
         $response = Http::withBasicAuth($this->userAuth(), $this->passAuth())
             ->get($this->connection());
 
         return $response->json();
-    }
-
-    /**
-     * Perform API GET.
-     * Returns the list of all the Live calls.
-     *
-     * @return json response
-     */
-    public function showliveCalls()
-    {
-        return $this->liveCalls();
     }
 }
